update: realTimeClock proj
