+++ src/App.js
@@ -1,7 +1,7 @@
 import { Routes, Route } from 'react-router-dom';
 import Homepage from './pages/Homepage';
 import Projects from './pages/Projects';
-import Clock from './pages/Projects/RealTimeClock';
+import RealTimeClock from './pages/Projects/RealTimeClock';
 
 function App() {
   return (
@@ -11,7 +11,7 @@ function App() {
         <Route path="/products" element={<h1>Hi</h1>} />
         <Route path="/about" element={<h1>About</h1>} />
         <Route path="/projects" element={<Projects />} />
-        <Route path="/projects/realTimeClock" element={<Clock />} />
+        <Route path="/projects/realTimeClock" element={<RealTimeClock />} />
       </Routes>
     </>
   );
