+++ src/pages/Projects/RealTimeClock/Clock/Hours/index.jsx
@@ -2,7 +2,7 @@ import React from "react";
 
 const Hour = ({ hours }) => {
   console.log("hours rendered");
-  return <h1 className="time"> {hours} </h1>;
+  return <h1 className="time"> {hours.toString().padStart(2, '0')} </h1>;
 };
 
 export default React.memo(Hour);
