+++ src/pages/Projects/RealTimeClock/Clock/Minutes/index.jsx
@@ -2,7 +2,7 @@ import React from "react";
 
 const Minute = ({ min }) => {
   console.log("min rendered");
-  return <h1 className="time"> {min} </h1>;
+  return <h1 className="time"> {min.toString().padStart(2, '0')} </h1>;
 };
 
 export default React.memo(Minute);
