+++ src/pages/Projects/RealTimeClock/Clock/Seconds/index.jsx
@@ -2,7 +2,7 @@ import React from "react";
 
 const Seconds = ({ sec }) => {
   console.log("sec rendered");
-  return <h1 className="time"> {sec} </h1>;
+  return <h1 className="time"> {sec.toString().padStart(2, '0')} </h1>;
 };
 
 export default React.memo(Seconds);
